LPS-105380 Fix violations
