+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/ui/AccountGroupAccountEntriesScreenNavigationCategory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
+package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib.ui;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
 import com.liferay.account.admin.web.internal.display.AccountGroupDisplay;
@@ -20,6 +20,7 @@ import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.User;
 
 import java.io.IOException;
 
@@ -36,34 +37,45 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	property = {
-		"screen.navigation.category.order:Integer=10",
+		"screen.navigation.category.order:Integer=20",
 		"screen.navigation.entry.order:Integer=10"
 	},
 	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
 )
-public class AccountGroupInformationScreenNavigationCategory
+public class AccountGroupAccountEntriesScreenNavigationCategory
 	implements ScreenNavigationCategory,
 			   ScreenNavigationEntry<AccountGroupDisplay> {
 
 	@Override
 	public String getCategoryKey() {
-		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_INFORMATION;
+		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_ACCOUNTS;
 	}
 
 	@Override
 	public String getEntryKey() {
-		return AccountScreenNavigationEntryConstants.ENTRY_KEY_INFORMATION;
+		return AccountScreenNavigationEntryConstants.ENTRY_KEY_ACCOUNTS;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "information");
+		return LanguageUtil.get(locale, "accounts");
 	}
 
 	@Override
 	public String getScreenNavigationKey() {
 		return AccountScreenNavigationEntryConstants.
 			SCREEN_NAVIGATION_KEY_ACCOUNT_GROUP;
+	}
+
+	@Override
+	public boolean isVisible(
+		User user, AccountGroupDisplay accountGroupDisplay) {
+
+		if (accountGroupDisplay.getAccountGroupId() == 0) {
+			return false;
+		}
+
+		return true;
 	}
 
 	@Override
@@ -74,7 +86,7 @@ public class AccountGroupInformationScreenNavigationCategory
 
 		jspRenderer.renderJSP(
 			httpServletRequest, httpServletResponse,
-			"/account_groups_admin/account_group/information.jsp");
+			"/account_groups_admin/account_group/view_account_entries.jsp");
 	}
 
 	@Reference
